fix some main styles
