+++ src/pages/FavoritePage/FavoritePage.module.css
@@ -1,4 +1,4 @@
-.favoritesContainer {
+/* .favoritesContainer {
   max-width: 375px;
   min-height: 800px;
   position: relative;
@@ -14,4 +14,4 @@
     padding-left: 100px;
     padding-right: 100px;
   }
-}
+} */
