+++ src/pages/MyRecipesPage/MyRecipesPage.js
@@ -7,7 +7,7 @@ import { fetchMyRecipes, deleteMyRecipes } from 'redux/MyRecipe/MyRecipeOperatio
 import { useEffect } from 'react';
 import { Paginator } from 'components/Paginator/Paginator';
 import { changePage } from 'redux/MyRecipe/MyRecipeSlice';
-import css from "./MyRecipesPage.module.css";
+// import css from "./MyRecipesPage.module.css";
 import { NoRecipe } from 'components/NoRecipe/NoRecipe';
 
 export default function MyRecipesPage() {
@@ -19,7 +19,7 @@ export default function MyRecipesPage() {
     dispatch(fetchMyRecipes(page));
   }, [dispatch, page]);
   return (
-    <section className={css.myRecipeContainer}>
+    <section>
       <MainTitle title="My recipes" />
       {recipes.data.length > 0 ? (
         <>
