+++ src/pages/MyRecipesPage/MyRecipesPage.module.css
@@ -1,15 +1,9 @@
-.myRecipeContainer {
-  max-width: 375px;
-  margin: 0 auto;
-  padding: 80px 20px;
+/* .myRecipeContainer {
+  padding: 80px 0;
   @media screen and (min-width: 768px) {
-    max-width: 768px;
-    padding: 140px 32px;
+    padding: 140px 0;
   }
   @media screen and (min-width: 1440px) {
-    max-width: 1440px;
     padding-top: 158px;
-    padding-left: 100px;
-    padding-right: 100px;
   }
-}
+} */
